optimized InputText component
