+++ src/components/ModalCreate.tsx
@@ -1,4 +1,4 @@
-import InputText from "./InputText";
+import InputText from "./input-text/InputText";
 import CheckboxImage from "./CheckboxImage";
 import { Categories } from "@/interfaces/Categories";
 import { cubeCollection } from "@/lib/cubeCollection";
@@ -100,7 +100,9 @@ export default function ModalCreate() {
         tabIndex={2}
         aria-hidden={false}
         className={`fixed backdrop-blur-[2px] top-0 left-0 z-50 w-full p-4 overflow-x-hidden overflow-y-auto md:inset-0 h-screen flex flex-col items-center`}
-        onClick={(e) => {if(e.target === e.currentTarget) setModalOpen(false)}}
+        onClick={(e) => {
+          if (e.target === e.currentTarget) setModalOpen(false);
+        }}
       >
         <div className="relative w-full max-w-2xl max-h-full">
           {/* <!-- Modal content --> */}
